requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
